feat: Using new even API for notifying workers about disconnection
